+++ website/docs/table-group/data-table/examples/column-alignment.tsx
@@ -3,7 +3,7 @@ import DataTable from 'intergalactic/data-table';
 
 const Demo = () => {
   return (
-    <DataTable data={data}>
+    <DataTable data={data} aria-label={'Table title. Column alignment'}>
       <DataTable.Head>
         <DataTable.Column name='keyword' children='Keyword' />
         <DataTable.Column name='kd' children='KD,%' justifyContent='flex-end' />