+++ website/docs/table-group/data-table/examples/scroll-in-table.tsx
@@ -3,7 +3,7 @@ import DataTable from 'intergalactic/data-table';
 
 const Demo = () => {
   return (
-    <DataTable data={data}>
+    <DataTable data={data} aria-label={'Table title. Scroll inside'}>
       <DataTable.Head wMin={1000}>
         <DataTable.Column name='keyword' children='Keyword' />
         <DataTable.Column name='kd' children='KD,%' />